add an overload of 'PrintPolynomial' taking a smart pointer
This will avoid some noise in an upcoming change of having to insert `.get()` at
every call site of `PrintPolynomial`.

This is somewhat unorthodox. Usually functions would take _either_ an owning
reference as a `std::shared_ptr` (forcing a copy and a reference count
increment) or a non-owning reference as a raw pointer. `PrintPolynomial` is a
morally const operation on the polynomial and would normally take it as a raw
pointer. But it calls `GetTotalTermCount` that modifies some cached information
in the `XPolynomial`. We could describe this by having it take a const raw
pointer and qualify the members of `XPolynomial` that are modified by
`GetTotalTermCount` as `mutable`. But it seems less complex and error prone to
instead add this unorthodox alternative of taking a smart pointer by reference.